+++ Framework/Sources/SpreadsheetView+Layout.swift
@@ -9,6 +9,64 @@
 import UIKit
 
 extension SpreadsheetView {
+    public override func layoutSubviews() {
+        super.layoutSubviews()
+
+        tableView.delegate = nil
+        columnHeaderView.delegate = nil
+        rowHeaderView.delegate = nil
+        cornerView.delegate = nil
+        defer {
+            tableView.delegate = self
+            columnHeaderView.delegate = self
+            rowHeaderView.delegate = self
+            cornerView.delegate = self
+
+            needsReload = false
+        }
+
+        reloadDataIfNeeded()
+
+        guard numberOfColumns > 0 && numberOfRows > 0 else {
+            return
+        }
+
+        layoutCorner()
+        layoutRowHeader()
+        layoutColumnHeader()
+
+        if needsReload {
+            adjustScrollViewFrames()
+
+            if circularScrollingOptions.direction.contains(.horizontally) {
+                scrollToHorizontalCenter()
+                if circularScrollingOptions.headerStyle == .rowHeaderStartsFirstColumn {
+                    layoutRowHeader()
+                }
+            }
+            if circularScrollingOptions.direction.contains(.vertically) {
+                scrollToVerticalCenter()
+                if circularScrollingOptions.headerStyle == .columnHeaderStartsFirstRow {
+                    layoutColumnHeader()
+                }
+            }
+        }
+
+        layoutTable()
+
+        if needsReload {
+            adjustOverlayViewContentSize()
+            arrangeScrollViews()
+        }
+
+        if circularScrollingOptions.direction.contains(.horizontally) {
+            recenterHorizontallyIfNecessary()
+        }
+        if circularScrollingOptions.direction.contains(.vertically) {
+            recenterVerticallyIfNecessary()
+        }
+    }
+
     func layout(scrollView: ScrollView) {
         let layoutEngine = LayoutEngine(spreadsheetView: self, scrollView: scrollView)
         layoutEngine.layout()
@@ -221,22 +279,26 @@ extension SpreadsheetView {
                 columnHeaderView.frame.origin.y = frozenRows > 0 ? rowHeaderView.frame.height : 0
             }
 
-            let height = rootView.frame.height - rootView.contentInset.top - rootView.contentInset.bottom - (circularScrollingOptions.headerStyle == .columnHeaderStartsFirstRow ? 0 : rowHeaderView.frame.height)
+            let height = rootView.frame.height - (rootView.contentInset.top + rootView.contentInset.bottom) - (circularScrollingOptions.headerStyle == .columnHeaderStartsFirstRow ? 0 : rowHeaderView.frame.height)
             columnHeaderView.frame.size.height = height < 0 ? 0 : height
 
             tableView.frame.origin.x = columnHeaderView.frame.width - intercellSpacing.width
             tableView.frame.size.width = (rootView.frame.width - (rootView.contentInset.left + rootView.contentInset.right)) - (columnHeaderView.frame.width - intercellSpacing.width)
+        } else {
+            tableView.frame.size.width = (rootView.frame.width - (rootView.contentInset.left + rootView.contentInset.right))
         }
         if frozenRows > 0 {
             if circularScrollingOptions.headerStyle != .rowHeaderStartsFirstColumn {
                 rowHeaderView.frame.origin.x = frozenColumns > 0 ? columnHeaderView.frame.width : 0
             }
 
-            let width = rootView.frame.width - rootView.contentInset.left - rootView.contentInset.right - (circularScrollingOptions.headerStyle == .rowHeaderStartsFirstColumn ? 0 : columnHeaderView.frame.width)
+            let width = rootView.frame.width - (rootView.contentInset.left + rootView.contentInset.right) - (circularScrollingOptions.headerStyle == .rowHeaderStartsFirstColumn ? 0 : columnHeaderView.frame.width)
             rowHeaderView.frame.size.width = width < 0 ? 0 : width
 
             tableView.frame.origin.y = rowHeaderView.frame.height - intercellSpacing.height
             tableView.frame.size.height = (rootView.frame.height - (rootView.contentInset.top + rootView.contentInset.bottom)) - (rowHeaderView.frame.height - intercellSpacing.height)
+        } else {
+            tableView.frame.size.height = (rootView.frame.height - (rootView.contentInset.top + rootView.contentInset.bottom))
         }
         if frozenColumns > 0 && frozenRows > 0 {
             if circularScrollingOptions.headerStyle != .columnHeaderStartsFirstRow {
@@ -255,9 +317,11 @@ extension SpreadsheetView {
             + (frozenColumns > 0 ? -columnHeaderView.frame.width + intercellSpacing.width : 0)
         if width > 0 {
             if width != tableView.frame.size.width {
+                rowHeaderView.frame.size.width = width
                 tableView.frame.size.width = width
             }
         } else {
+            rowHeaderView.frame.size.width = 0
             tableView.frame.size.width = 0
         }
 
@@ -274,10 +338,9 @@ extension SpreadsheetView {
         }
     }
 
-    func adjustOverlayViewFrame() {
-        overlayView.scrollIndicatorInsets = rootView.scrollIndicatorInsets
-        overlayView.contentSize = CGSize(width: rootView.contentInset.left + rootView.contentInset.right + columnHeaderView.contentSize.width + tableView.contentSize.width,
-                                         height: rootView.contentInset.top + rootView.contentInset.bottom + rowHeaderView.contentSize.height + tableView.contentSize.height)
+    func adjustOverlayViewContentSize() {
+        overlayView.contentSize = CGSize(width: rootView.contentInset.left + rootView.contentInset.right + tableView.frame.origin.x - intercellSpacing.width + tableView.contentSize.width,
+                                         height: rootView.contentInset.top + rootView.contentInset.bottom + tableView.frame.origin.y - intercellSpacing.height + tableView.contentSize.height)
     }
 
     func arrangeScrollViews() {