+++ Framework/Sources/SpreadsheetView.swift
@@ -9,17 +9,36 @@
 import UIKit
 
 public class SpreadsheetView: UIView {
+    /// The object that provides the data for the collection view.
+    ///
+    /// - Note: The data source must adopt the `SpreadsheetViewDataSource` protocol.
+    ///   The spreadsheet view maintains a weak reference to the data source object.
     public weak var dataSource: SpreadsheetViewDataSource? {
         didSet {
             resetTouchHandlers(to: [tableView, columnHeaderView, rowHeaderView, cornerView])
             reloadData()
         }
     }
+    /// The object that acts as the delegate of the spreadsheet view.
+    /// - Note: The delegate must adopt the `SpreadsheetViewDelegate` protocol.
+    ///   The spreadsheet view maintains a weak reference to the delegate object.
+    ///
+    ///   The delegate object is responsible for managing selection behavior and interactions with individual items.
     public weak var delegate: SpreadsheetViewDelegate?
 
+    /// The horizontal and vertical spacing between cells.
+    /// 
+    /// - Note: The default spacing is `(1.0, 1.0)`. Negative values are not supported.
     public var intercellSpacing = CGSize(width: 1, height: 1)
     public var gridStyle: GridStyle = .solid(width: 1, color: .lightGray)
 
+    /// A Boolean value that indicates whether users can select cells in the spreadsheet view.
+    ///
+    /// - Note: If the value of this property is `true` (the default), users can select cells.
+    ///   If you want more fine-grained control over the selection of cells,
+    ///   you must provide a delegate object and implement the appropriate methods of the `SpreadsheetViewDelegate` protocol.
+    ///
+    /// - SeeAlso: `allowsMultipleSelection`
     public var allowsSelection = true {
         didSet {
             if !allowsSelection {
@@ -27,6 +46,15 @@ public class SpreadsheetView: UIView {
             }
         }
     }
+    /// A Boolean value that determines whether users can select more than one cell in the spreadsheet view.
+    ///
+    /// - Note: This property controls whether multiple cells can be selected simultaneously.
+    ///   The default value of this property is `false`.
+    ///
+    ///   When the value of this property is true, tapping a cell adds it to the current selection (assuming the delegate permits the cell to be selected).
+    ///   Tapping the cell again removes it from the selection.
+    ///
+    /// - SeeAlso: `allowsSelection`
     public var allowsMultipleSelection = false {
         didSet {
             if allowsMultipleSelection {
@@ -35,16 +63,35 @@ public class SpreadsheetView: UIView {
         }
     }
 
+    /// A Boolean value that controls whether the vertical scroll indicator is visible.
+    ///
+    /// The default value is `true`. The indicator is visible while tracking is underway and fades out after tracking.
     public var showsVerticalScrollIndicator = true {
         didSet {
             overlayView.showsVerticalScrollIndicator = showsVerticalScrollIndicator
         }
     }
+    /// A Boolean value that controls whether the horizontal scroll indicator is visible.
+    ///
+    /// The default value is `true`. The indicator is visible while tracking is underway and fades out after tracking.
     public var showsHorizontalScrollIndicator = true {
         didSet {
             overlayView.showsHorizontalScrollIndicator = showsHorizontalScrollIndicator
         }
     }
+
+    /// A Boolean value that controls whether the scroll-to-top gesture is enabled.
+    ///
+    /// - Note: The scroll-to-top gesture is a tap on the status bar. When a user makes this gesture,
+    /// the system asks the scroll view closest to the status bar to scroll to the top.
+    /// If that scroll view has `scrollsToTop` set to `false`, its delegate returns false from `scrollViewShouldScrollToTop(_:)`, 
+    /// or the content is already at the top, nothing happens.
+    ///
+    /// After the scroll view scrolls to the top of the content view, it sends the delegate a `scrollViewDidScrollToTop(_:)` message.
+    ///
+    /// The default value of scrollsToTop is `true`.
+    ///
+    /// On iPhone, the scroll-to-top gesture has no effect if there is more than one scroll view on-screen that has `scrollsToTop` set to `true`.
     public var scrollsToTop: Bool = true {
         didSet {
             tableView.scrollsToTop = scrollsToTop
@@ -67,6 +114,12 @@ public class SpreadsheetView: UIView {
     var circularScrollScalingFactor: (horizontal: Int, vertical: Int) = (1, 1)
     var centerOffset = CGPoint.zero
 
+    /// The view that provides the background appearance.
+    ///
+    /// - Note: The view (if any) in this property is positioned underneath all of the other content and sized automatically to fill the entire bounds of the spreadsheet view.
+    /// The background view does not scroll with the spreadsheet view’s other content. The spreadsheet view maintains a strong reference to the background view object.
+    ///
+    /// This property is nil by default, which displays the background color of the spreadsheet view.
     public var backgroundView: UIView? {
         willSet {
             backgroundView?.removeFromSuperview()
@@ -75,17 +128,28 @@ public class SpreadsheetView: UIView {
             if let backgroundView = backgroundView {
                 backgroundView.frame = bounds
                 backgroundView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-                insertSubview(backgroundView, at: 0)
+                super.insertSubview(backgroundView, at: 0)
             }
         }
     }
 
+    /// Returns an array of visible cells currently displayed by the spreadsheet view.
+    ///
+    /// - Note: This method returns the complete list of visible cells displayed by the collection view.
+    ///
+    /// - Returns: An array of `Cell` objects. If no cells are visible, this method returns an empty array.
     public var visibleCells: [Cell] {
-        let cells: [Cell] = Array(columnHeaderView.visibleCells.values) + Array(rowHeaderView.visibleCells.values)
-            + Array(cornerView.visibleCells.values) + Array(tableView.visibleCells.values)
+        let cells: [Cell] = Array(columnHeaderView.visibleCells) + Array(rowHeaderView.visibleCells)
+            + Array(cornerView.visibleCells) + Array(tableView.visibleCells)
         return cells.sorted()
     }
 
+
+    /// An array of the visible items in the collection view.
+    /// - Note: The value of this property is a sorted array of IndexPath objects, each of which corresponds to a visible cell in the spreadsheet view.
+    /// If there are no visible items, the value of this property is an empty array.
+    ///
+    /// - SeeAlso: `visibleCells`
     public var indexPathsForVisibleItems: [IndexPath] {
         return visibleCells.map { $0.indexPath }
     }
@@ -94,16 +158,31 @@ public class SpreadsheetView: UIView {
         return Array(selectedIndexPaths).sorted().first
     }
 
+    /// The index paths for the selected items.
+    /// - Note: The value of this property is an array of IndexPath objects, each of which corresponds to a single selected item.
+    /// If there are no selected items, the value of this property is nil.
     public var indexPathsForSelectedItems: [IndexPath] {
         return Array(selectedIndexPaths).sorted()
     }
 
+    /// A Boolean value that determines whether scrolling is disabled in a particular direction.
+    /// - Note: If this property is `false`, scrolling is permitted in both horizontal and vertical directions.
+    /// If this property is `true` and the user begins dragging in one general direction (horizontally or vertically), the scroll view disables scrolling in the other direction.
+    /// If the drag direction is diagonal, then scrolling will not be locked and the user can drag in any direction until the drag completes.
+    /// The default value is `false`
     public var isDirectionalLockEnabled = false {
         didSet {
             tableView.isDirectionalLockEnabled = isDirectionalLockEnabled
         }
     }
 
+    /// A Boolean value that controls whether the scroll view bounces past the edge of content and back again.
+    /// - Note: If the value of this property is `true`, the scroll view bounces when it encounters a boundary of the content.
+    /// Bouncing visually indicates that scrolling has reached an edge of the content.
+    /// If the value is `false`, scrolling stops immediately at the content boundary without bouncing.
+    /// The default value is `true`.
+    ///
+    /// - SeeAlso: `alwaysBounceHorizontal`, `alwaysBounceVertical`
     public var bounces: Bool {
         get {
             return tableView.bounces
@@ -112,6 +191,12 @@ public class SpreadsheetView: UIView {
             tableView.bounces = newValue
         }
     }
+
+    /// A Boolean value that determines whether bouncing always occurs when vertical scrolling reaches the end of the content.
+    /// - Note: If this property is set to true and `bounces` is `true`, vertical dragging is allowed even if the content is smaller than the bounds of the scroll view.
+    /// The default value is `false`.
+    ///
+    /// - SeeAlso: `alwaysBounceHorizontal`
     public var alwaysBounceVertical: Bool {
         get {
             return tableView.alwaysBounceVertical
@@ -120,6 +205,12 @@ public class SpreadsheetView: UIView {
             tableView.alwaysBounceVertical = newValue
         }
     }
+
+    /// A Boolean value that determines whether bouncing always occurs when horizontal scrolling reaches the end of the content view.
+    /// - Note: If this property is set to `true` and `bounces` is `true`, horizontal dragging is allowed even if the content is smaller than the bounds of the scroll view.
+    /// The default value is `false`.
+    ///
+    /// - SeeAlso: `alwaysBounceVertical`
     public var alwaysBounceHorizontal: Bool {
         get {
             return tableView.alwaysBounceHorizontal
@@ -129,6 +220,9 @@ public class SpreadsheetView: UIView {
         }
     }
 
+    /// A Boolean value that determines whether paging is enabled for the scroll view.
+    /// - Note: If the value of this property is `true`, the scroll view stops on multiples of the scroll view’s bounds when the user scrolls.
+    /// The default value is false.
     public var isPagingEnabled: Bool {
         get {
             return tableView.isPagingEnabled
@@ -137,15 +231,23 @@ public class SpreadsheetView: UIView {
             tableView.isPagingEnabled = newValue
         }
     }
+
+    /// A Boolean value that determines whether scrolling is enabled.
+    /// - Note: If the value of this property is `true`, scrolling is enabled, and if it is `false`, scrolling is disabled. The default is `true`.
+    ///
+    /// When scrolling is disabled, the scroll view does not accept touch events; it forwards them up the responder chain.
     public var isScrollEnabled: Bool {
         get {
             return tableView.isScrollEnabled
         }
         set {
             tableView.isScrollEnabled = newValue
+            overlayView.isScrollEnabled = newValue
         }
     }
 
+    /// The style of the scroll indicators.
+    /// - Note: The default style is `default`. See `UIScrollViewIndicatorStyle` for descriptions of these constants.
     public var indicatorStyle: UIScrollViewIndicatorStyle {
         get {
             return overlayView.indicatorStyle
@@ -155,6 +257,8 @@ public class SpreadsheetView: UIView {
         }
     }
 
+    /// A floating-point value that determines the rate of deceleration after the user lifts their finger.
+    /// - Note: Your application can use the `UIScrollViewDecelerationRateNormal` and UIScrollViewDecelerationRateFast` constants as reference points for reasonable deceleration rates.
     public var decelerationRate: CGFloat {
         get {
             return tableView.decelerationRate
@@ -191,8 +295,12 @@ public class SpreadsheetView: UIView {
 
     var cellClasses = [String: Cell.Type]()
     var cellNibs = [String: UINib]()
-    var cellReuseQueues = [String: CellReuseQueue]()
+    var cellReuseQueues = [String: ReuseQueue<Cell>]()
     let blankCellReuseIdentifier = UUID().uuidString
+
+    var horizontalGridlineReuseQueue = ReuseQueue<Gridline>()
+    var verticalGridlineReuseQueue = ReuseQueue<Gridline>()
+    var borderReuseQueue = ReuseQueue<Border>()
 
     var highlightedIndexPaths = Set<IndexPath>()
     var selectedIndexPaths = Set<IndexPath>()
@@ -217,7 +325,7 @@ public class SpreadsheetView: UIView {
         rootView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
         rootView.showsHorizontalScrollIndicator = false
         rootView.showsVerticalScrollIndicator = false
-        addSubview(rootView)
+        super.addSubview(rootView)
 
         tableView.frame = bounds
         tableView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
@@ -257,7 +365,7 @@ public class SpreadsheetView: UIView {
         rootView.addSubview(columnHeaderView)
         rootView.addSubview(rowHeaderView)
         rootView.addSubview(cornerView)
-        addSubview(overlayView)
+        super.addSubview(overlayView)
 
         [tableView, columnHeaderView, rowHeaderView, cornerView, overlayView].forEach {
             addGestureRecognizer($0.panGestureRecognizer)
@@ -291,7 +399,7 @@ public class SpreadsheetView: UIView {
 
     func refreshData() {
         layoutProperties = resetLayoutProperties()
-        circularScrollScalingFactor =  determineCircularScrollScalingFactor()
+        circularScrollScalingFactor = determineCircularScrollScalingFactor()
         centerOffset = calculateCenterOffset()
 
         cornerView.layoutAttributes = layoutAttributeForCornerView()
@@ -299,43 +407,53 @@ public class SpreadsheetView: UIView {
         rowHeaderView.layoutAttributes = layoutAttributeForRowHeaderView()
         tableView.layoutAttributes = layoutAttributeForTableView()
 
+        cornerView.resetReusableObjects()
+        columnHeaderView.resetReusableObjects()
+        rowHeaderView.resetReusableObjects()
+        tableView.resetReusableObjects()
+
         resetContentSize(of: cornerView)
         resetContentSize(of: columnHeaderView)
         resetContentSize(of: rowHeaderView)
         resetContentSize(of: tableView)
 
+        cornerView.frame = .zero
+        columnHeaderView.frame = CGRect(x: 0, y: 0, width: 0, height: rootView.frame.height)
+        rowHeaderView.frame = CGRect(x: 0, y: 0, width: rootView.frame.width, height: 0)
+        tableView.frame = CGRect(origin: .zero, size: rootView.frame.size)
+
         cornerView.frame.size = cornerView.contentSize
         columnHeaderView.frame.size.width = columnHeaderView.contentSize.width
         rowHeaderView.frame.size.height = rowHeaderView.contentSize.height
+        if frozenColumns > 0 && frozenRows > 0 {
+            columnHeaderView.frame.size.height -= cornerView.frame.height - intercellSpacing.height
+            rowHeaderView.frame.size.width -= cornerView.frame.width - intercellSpacing.width
+        }
     }
 
     public func dequeueReusableCell(withReuseIdentifier identifier: String, for indexPath: IndexPath) -> Cell {
-        if let cellReuseQueue = cellReuseQueues[identifier] {
-            if let cell = cellReuseQueue.dequeue() {
+        if let reuseQueue = cellReuseQueues[identifier] {
+            if let cell = reuseQueue.dequeue() {
                 cell.prepareForReuse()
-
                 return cell
             }
         } else {
-            let cellReuseQueue = CellReuseQueue()
-            cellReuseQueues[identifier] = cellReuseQueue
+            let reuseQueue = ReuseQueue<Cell>()
+            cellReuseQueues[identifier] = reuseQueue
         }
         if identifier == blankCellReuseIdentifier {
             let cell = BlankCell()
             cell.reuseIdentifier = identifier
-
             return cell
         }
         if let clazz = cellClasses[identifier] {
             let cell = clazz.init()
             cell.reuseIdentifier = identifier
-
             return cell
         }
         if let nib = cellNibs[identifier] {
             if let cell = nib.instantiate(withOwner: nil, options: nil).first as? Cell {
                 cell.reuseIdentifier = identifier
-
                 return cell
             }
         }
@@ -570,25 +688,25 @@ public class SpreadsheetView: UIView {
     }
 
     public func cellForItem(at indexPath: IndexPath) -> Cell? {
-        if let cell = tableView.visibleCells
+        if let cell = tableView.visibleCells.pairs
             .filter({ $0.key.row == indexPath.row && $0.key.column == indexPath.column })
             .map({ return $1 })
             .first {
             return cell
         }
-        if let cell = rowHeaderView.visibleCells
+        if let cell = rowHeaderView.visibleCells.pairs
             .filter({ $0.key.row == indexPath.row && $0.key.column == indexPath.column })
             .map({ return $1 })
             .first {
             return cell
         }
-        if let cell = columnHeaderView.visibleCells
+        if let cell = columnHeaderView.visibleCells.pairs
             .filter({ $0.key.row == indexPath.row && $0.key.column == indexPath.column })
             .map({ return $1 })
             .first {
             return cell
         }
-        if let cell = cornerView.visibleCells
+        if let cell = cornerView.visibleCells.pairs
             .filter({ $0.key.row == indexPath.row && $0.key.column == indexPath.column })
             .map({ return $1 })
             .first {
@@ -600,22 +718,22 @@ public class SpreadsheetView: UIView {
     public func cellsForItem(at indexPath: IndexPath) -> [Cell] {
         var cells = [Cell]()
         cells.append(contentsOf:
-            tableView.visibleCells
+            tableView.visibleCells.pairs
                 .filter { $0.key.row == indexPath.row && $0.key.column == indexPath.column }
                 .map { return $1 }
         )
         cells.append(contentsOf:
-            rowHeaderView.visibleCells
+            rowHeaderView.visibleCells.pairs
                 .filter { $0.key.row == indexPath.row && $0.key.column == indexPath.column }
                 .map { return $1 }
         )
         cells.append(contentsOf:
-            columnHeaderView.visibleCells
+            columnHeaderView.visibleCells.pairs
                 .filter { $0.key.row == indexPath.row && $0.key.column == indexPath.column }
                 .map { return $1 }
         )
         cells.append(contentsOf:
-            cornerView.visibleCells
+            cornerView.visibleCells.pairs
                 .filter { $0.key.row == indexPath.row && $0.key.column == indexPath.column }
                 .map { return $1 }
         )
@@ -631,90 +749,37 @@ public class SpreadsheetView: UIView {
         let columnRecords = columnHeaderView.columnRecords + tableView.columnRecords
         let rowRecords = rowHeaderView.rowRecords + tableView.rowRecords
 
-        let x = columnRecords[column] + (column >= frozenColumns ? tableView.frame.origin.x : 0) + intercellSpacing.width
-        let y = rowRecords[row] + (row >= frozenRows ? tableView.frame.origin.y : 0) + intercellSpacing.height
-        let origin = CGPoint(x: x, y: y)
-
+        let origin: CGPoint
         let size: CGSize
-        if let mergedCell = mergedCell(for: Location(row: row, column: column)) {
-            let width = (mergedCell.from.column...mergedCell.to.column).reduce(0) { $0 + layoutProperties.columnWidthCache[$1] }
-            let height = (mergedCell.from.row...mergedCell.to.row).reduce(0) { $0 + layoutProperties.rowHeightCache[$1] }
-            size = CGSize(width: width, height: height)
-        } else {
-            size = CGSize(width: layoutProperties.columnWidthCache[column], height: layoutProperties.rowHeightCache[row])
+        func originFor(column: Int, row: Int) -> CGPoint {
+            let x = columnRecords[column] + (column >= frozenColumns ? tableView.frame.origin.x : 0) + intercellSpacing.width
+            let y = rowRecords[row] + (row >= frozenRows ? tableView.frame.origin.y : 0) + intercellSpacing.height
+            return CGPoint(x: x, y: y)
         }
+        if let mergedCell = mergedCell(for: Location(row: row, column: column)) {
+            origin = originFor(column: mergedCell.from.column, row: mergedCell.from.row)
 
+            var width: CGFloat = 0
+            var height: CGFloat = 0
+            for column in mergedCell.from.column...mergedCell.to.column {
+                width += layoutProperties.columnWidthCache[column]
+            }
+            for row in mergedCell.from.row...mergedCell.to.row {
+                height += layoutProperties.rowHeightCache[row]
+            }
+            size = CGSize(width: width + intercellSpacing.width * CGFloat(mergedCell.columnCount - 1),
+                          height: height + intercellSpacing.height * CGFloat(mergedCell.rowCount - 1))
+        } else {
+            origin = originFor(column: column, row: row)
+
+            let width = layoutProperties.columnWidthCache[column]
+            let height = layoutProperties.rowHeightCache[row]
+            size = CGSize(width: width, height: height)
+        }
         return CGRect(origin: origin, size: size)
     }
 
     func mergedCell(for indexPath: Location) -> CellRange? {
         return layoutProperties.mergedCellLayouts[indexPath]
-    }
-
-    public override func layoutSubviews() {
-        super.layoutSubviews()
-
-        tableView.delegate = nil
-        columnHeaderView.delegate = nil
-        rowHeaderView.delegate = nil
-        cornerView.delegate = nil
-        defer {
-            tableView.delegate = self
-            columnHeaderView.delegate = self
-            rowHeaderView.delegate = self
-            cornerView.delegate = self
-
-            needsReload = false
-        }
-
-        reloadDataIfNeeded()
-
-        guard numberOfColumns > 0 && numberOfRows > 0 else {
-            return
-        }
-
-        layoutCorner()
-        if needsReload && frozenColumns > 0 && frozenRows > 0 {
-            columnHeaderView.frame.size.height -= cornerView.frame.height - intercellSpacing.height
-            rowHeaderView.frame.size.width -= cornerView.frame.width - intercellSpacing.width
-        }
-
-        layoutRowHeader()
-        layoutColumnHeader()
-
-        if needsReload {
-            adjustScrollViewFrames()
-
-            if circularScrollingOptions.direction.contains(.horizontally) {
-                scrollToHorizontalCenter()
-                if circularScrollingOptions.headerStyle == .rowHeaderStartsFirstColumn {
-                    layoutRowHeader()
-                }
-            }
-            if circularScrollingOptions.direction.contains(.vertically) {
-                scrollToVerticalCenter()
-                if circularScrollingOptions.headerStyle == .columnHeaderStartsFirstRow {
-                    layoutColumnHeader()
-                }
-            }
-        }
-
-        layoutTable()
-
-        if needsReload {
-            adjustOverlayViewFrame()
-            arrangeScrollViews()
-        }
-
-        if circularScrollingOptions.direction.contains(.horizontally) {
-            recenterHorizontallyIfNecessary()
-        }
-        if circularScrollingOptions.direction.contains(.vertically) {
-            recenterVerticallyIfNecessary()
-        }
-    }
-
-    public override func isKind(of aClass: AnyClass) -> Bool {
-        return rootView.isKind(of: aClass)
     }
 }